refactor: Consolidate CobaltWebContents{Delegatte/Observer} to Cobalt Shell
