+++ cobalt/BUILD.gn
@@ -268,9 +268,9 @@ test("cobalt_unittests") {
     "//cobalt/browser/constants:cobalt_experiment_names",
     "//cobalt/browser/h5vcc_metrics/public/mojom",
     "//cobalt/browser/h5vcc_runtime:deep_link_manager",
-    "//cobalt/browser/migrate_storage_record:unittest",
     "//cobalt/browser/user_agent",
     "//cobalt/common/libc:libc_test",
+    "//cobalt/shell/browser/migrate_storage_record:unittest",
     "//components/metrics",
     "//components/prefs",
     "//components/prefs:test_support",
