+++ cobalt/browser/BUILD.gn
@@ -28,8 +28,6 @@ source_set("browser") {
     "cobalt_content_browser_client.h",
     "cobalt_secure_navigation_throttle.cc",
     "cobalt_secure_navigation_throttle.h",
-    "cobalt_web_contents_delegate.cc",
-    "cobalt_web_contents_delegate.h",
     "cobalt_web_contents_observer.cc",
     "cobalt_web_contents_observer.h",
   ]
@@ -44,7 +42,6 @@ source_set("browser") {
   public_deps = [ "//content/public/browser" ]
 
   deps = [
-    ":embed_polyfilled_javascript",
     ":features",
     ":global_features",
     ":metrics",
@@ -62,7 +59,6 @@ source_set("browser") {
     "//cobalt/browser/h5vcc_runtime/public/mojom",
     "//cobalt/browser/h5vcc_system",
     "//cobalt/browser/h5vcc_system/public/mojom",
-    "//cobalt/browser/migrate_storage_record",
     "//cobalt/browser/performance",
     "//cobalt/browser/performance/public/mojom",
     "//cobalt/browser/user_agent",
@@ -71,7 +67,6 @@ source_set("browser") {
     "//cobalt/media/service/mojom",
     "//cobalt/shell:cobalt_shell_lib",
     "//components/embedder_support:user_agent",
-    "//components/js_injection/browser:browser",
     "//components/metrics",
     "//components/metrics:test_support",
     "//components/metrics_services_manager",
@@ -85,7 +80,6 @@ source_set("browser") {
     deps += [
       "//cobalt/android:browser_jni_headers",
       "//cobalt/android:jni_headers",
-      "//cobalt/android/oom_intervention",
     ]
   } else {
     deps += [ "//cobalt/browser/crash_annotator" ]
@@ -109,23 +103,6 @@ source_set("switches") {
   deps = [
     "//base",
     "//cobalt/shell:cobalt_shell_switches",
-  ]
-}
-
-config("embed_polyfilled_javascript_config") {
-  include_dirs = [ root_gen_dir ]
-}
-
-# TODO(cobalt, b/390710539): migrate to GRIT.
-action("embed_polyfilled_javascript") {
-  script = "//cobalt/build/generate_data_header.py"
-  outputs = [ "$target_gen_dir/embedded_resources/embedded_js.h" ]
-  public_configs = [ ":embed_polyfilled_javascript_config" ]
-
-  args = [
-    "CobaltJavaScriptPolyfill",
-    rebase_path(outputs[0], root_build_dir),
-    rebase_path("embedded_resources", root_build_dir),
   ]
 }
