+++ cobalt/shell/BUILD.gn
@@ -48,16 +48,19 @@ group("cobalt_shell_lib_deps") {
   public_deps = [
     ":android_shell_descriptors",
     ":embed_loader_resources_as_header_files",
+    ":embed_polyfilled_javascript",
     ":resources",
     "//base",
     "//base:base_static",
     "//cc/base",
+    "//cobalt/shell/browser/migrate_storage_record",
     "//components/cdm/renderer",
     "//components/crash/content/browser",
     "//components/crash/core/app",
     "//components/custom_handlers",
     "//components/embedder_support:embedder_support",
     "//components/embedder_support:user_agent",
+    "//components/js_injection/browser:browser",
     "//components/keyed_service/content",
     "//components/metrics",
     "//components/metrics:net",
@@ -159,6 +162,10 @@ group("cobalt_shell_lib_deps") {
     if (is_android) {
       public_deps += [ "//starboard/android/shared:starboard_platform" ]
     }
+  }
+
+  if (is_androidtv) {
+    public_deps += [ "//cobalt/android/oom_intervention" ]
   }
 }
 
@@ -463,13 +470,29 @@ group("resources") {
 
 action("embed_loader_resources_as_header_files") {
   script = "//cobalt/build/generate_data_header.py"
-
-  inputs = [ "embedded_resources/splash.html" ]
   outputs = [ "$target_gen_dir/embedded_resources/embedded_resources.h" ]
 
   args = [
     "LoaderEmbeddedResources",
     rebase_path(outputs[0], root_build_dir),
-    rebase_path("embedded_resources", root_build_dir),
+    rebase_path("embedded_resources/loader_embedded_resources", root_build_dir),
+  ]
+}
+
+config("embed_polyfilled_javascript_config") {
+  include_dirs = [ root_gen_dir ]
+}
+
+# TODO(cobalt, b/390710539): migrate to GRIT.
+action("embed_polyfilled_javascript") {
+  script = "//cobalt/build/generate_data_header.py"
+  outputs = [ "$target_gen_dir/embedded_resources/embedded_js.h" ]
+  public_configs = [ ":embed_polyfilled_javascript_config" ]
+
+  args = [
+    "CobaltJavaScriptPolyfill",
+    rebase_path(outputs[0], root_build_dir),
+    rebase_path("embedded_resources/cobalt_java_script_polyfill",
+                root_build_dir),
   ]
 }
