+++ cobalt/shell/browser/migrate_storage_record/storage.proto
